backend/requirements.txt
pytest>=7.0
google-genai>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pypdf>=4.0.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.4
